WIP: Update MRI upstreams for Phosphorus
Adopt latest versions, namely;
- odlparent-9.0.2
- infrautils-2.0.2
- mdsal-8.0.0
- controller-4.0.0
- aaa-0.14.0-SNAPSHOT
- netconf-2.0.0-SNAPSHOT

Also introduce model-{inventory,topology}, which were removed from
upstream.

Since these versions also force the Uint8/Uint32 vs. Short/Long
equivalence issue, we mass-migrate all APIs to use Uint constructs.
This in turn requires minor adjustments to mocks, as we no longer
get automatic smart zeroes.

Another thing that is cropping up here is that augment-identifier
is now reflected in package names properly, requiring a proper update
of imports and references.

+++ extension/openflowjava-extension-nicira/src/test/java/org/opendaylight/openflowjava/nx/NiciraExtensionCodecRegistratorImplTest.java
@@ -38,6 +38,7 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.common.action.rev1
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.oxm.rev150225.MatchField;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.oxm.rev150225.Nxm0Class;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.oxm.rev150225.match.entries.grouping.MatchEntry;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 @RunWith(MockitoJUnitRunner.class)
 public class NiciraExtensionCodecRegistratorImplTest {
@@ -57,7 +58,7 @@ public class NiciraExtensionCodecRegistratorImplTest {
     @Mock
     OFDeserializer<MatchEntry> matchDeserializer;
 
-    public static final short VERSION = 4;
+    public static final Uint8 VERSION = Uint8.valueOf(4);
     public static final byte VERSION1 = EncodeConstants.OF10_VERSION_ID;
     public static final byte VERSION2 = EncodeConstants.OF13_VERSION_ID;
